+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Versions
 ## 3.8.0
 HOLIDAYS
-- Add Hallows End (v1) for Alliance
+- Add Hallows End (v1) full guide for Alliance
+- Add Hallows End (v1) stub for Horde
 
 ## 3.7.0
 MOLTEN FRONT